CellBase.jl

Documentation for CellBase.jl

What is package does

Provide a basic low level interface for storing and handling crystal structure data. The target application is for small and periodic cells, with the emphasis on both ease to use as well as performance.

Guides

Lattice type

The Lattice type represents the lattice. The lattice vectors are stored as a matrix of column vectors. For example, the first lattice vector should be accessed as cellmat(lattice)[:, 1].

Updating an existing `Lattice`

The reciprocal lattice is also stored for quick access, however, this means that whenever the matrix is directly modified, the store reciprocal lattice vectors should be updated as well. Hence, the set_cellmat! function should be used for updating the cell matrix, which does this is automatically.

CellBase.Lattice — Type

Lattice{T}

The type represents a lattice in 3D space formed by three column lattice vectors.

CellBase.Lattice — Method

Lattice(matrix::Matrix{T}) where T

Construct a Lattice from a matrix of column vectors.

CellBase.Lattice — Method

Lattice(a::T, b::T, c::T, α::T, β::T, γ::T) where T

Construct a Lattice from lattice parameters.

CellBase.Lattice — Method

Lattice(a::T, b::T, c::T) where T <: Real

Construct a Lattice with orthogonal lattice vectors.

CellBase.Lattice — Method

Lattice(va::Vector{T}, vb::Vector{T}, vc::Vector{T}) where T

Construct a Lattice from three lattice vectors.

CellBase.Lattice — Method

Lattice(cellpar::Vector{T}) where T

Construct a Lattice from a six-vector of lattice parameters.

CellBase.cellmat — Method

cellmat(lattice::Lattice)

Return the matrix of column vectors.

CellBase.cellmat_row — Method

Get a matrix of column vectors

CellBase.cellpar — Method

cellpar(lattice::Lattice)

Return the lattice parameters as a six-vector.

CellBase.cellvecs — Method

Lattice vectors

CellBase.frac_pos — Method

Fraction positions of a site

CellBase.get_cellmat — Method

Return a copy of the cell matrix

CellBase.get_rec_cellmat — Method

Return a copy of the reciprocal cell matrix

CellBase.get_scaled_positions — Method

get_scaled_positions(l::Lattice, v::AbstractVecOrMat) = reciprocal(l) * v

Return the scaled positions.

CellBase.mic — Method

mic(l::Lattice, vec::AbstractVecOrMat)

Compute the minimum-image convention representation of a series of displacement vectors.

CellBase.mic_naive — Method

Compute naive MIC representation of the vector(s)

Not safe for skewed cells. Requires

norm(length) < 0.5 * min(cellpar(l)[1:3])

See also:

W. Smith, "The Minimum Image Convention in Non-Cubic MD Cells", 1989, http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.57.1696.

CellBase.mic_naive — Method

 mic_safe(l::Lattice, v::AbstractVector)

Compute MIC representation of vectors using a safe approach based on Minkowski reduced cell

CellBase.mic_safe — Method

 mic_safe(l::Lattice, v::AbstractMatrix)

Compute MIC representation of vectors using a safe approach based on Minkowski reduced cell

CellBase.mic_safe — Method

 mic_safe(l::Lattice, v::AbstractVector)

Compute MIC representation of vectors using a safe approach based on Minkowski reduced cell

CellBase.random_vec — Method

Random displacement vector in a unit cell

CellBase.random_vec_in_cell — Method

random_vec_in_cell(cell::Matrix{T}; scales::Vector=ones(size(cell)[1])) where T

Get an random vector within a unit cell. The cell is a matrix made of column vectors.

CellBase.rec_cellmat — Method

rec_cellmat(l::Lattice)

Returns the matrix of reciprocal lattice vectors (without the $2\pi$ factor).

CellBase.reciprocal — Method

reciprocal(l::Lattice)

Returns the matrix of reciprocal lattice vectors (without the $2\pi$ factor).

CellBase.scaled_positions — Function

Alias to get_scaled_positions

CellBase.set_cellmat! — Method

set_cellmat!(lattice::Lattice, mat)

Set the cell matrix and update the reciprocal cell matrix.

CellBase.update_rec! — Method

update_rec!(l::Lattice)

Update the reciprocal matrix - this should be called everytime cell is changed.

CellBase.volume — Method

Return the volume of the cell

CellBase.wrap! — Method

Wrap a site back into the box

CellBase.wrap_positions — Method

wrap_positions(l::Lattice, v::AbstractVecOrMat)

Wrap positions back into the box defined by the cell vectors.

CellBase.max_shifts — Method

Compute the require periodicities required to fill a cut off sphere of a certain radius

CellBase.shift_indices — Method

shift_indices(lattice::AbstractArray, rc::Real)

Compute a vector containing shift indices

CellBase.shift_vectors — Function

shift_vectors(lattice::AbstractMatrix, shift1, shift2, shift3)

Compute the shift vectors needed to include all parts of the lattice within a cut off radius.

CellBase.shift_vectors — Method

Compute shift vectors given lattice matrix and cut off radius

Cell type

The Cell type represent a crystal structure combining the positions of atoms and the lattice vectors. It basically combines the Lattice, the atomic positions and their identities. Internally, the formere is stored as an matrix of column vectors of the absolute positions (cartesian space).

For flexibility, any additional array like data can be stored in the arrays property which is a dictionary with Symbol keys. Additional metadata can be also be stored under metadata, which is a Dict{Any, Any}.

CellBase.Cell — Type

A Cell represents a periodic structure in three-dimensional space.

Defined as:

mutable struct Cell{T}
    lattice::Lattice{T}                 # Lattice of the structure
    symbols::Vector{Symbol}
    positions::Matrix{T}
    arrays::Dict{Symbol, AbstractArray}        # Any additional arrays
    metadata::Dict
end

CellBase.Cell — Method

Cell(lat::Lattice, numbers::Vector{Int}, positions::Vector)

Constructure the Cell type from lattice, positions and numbers

CellBase.Cell — Method

Cell(l::Lattice, symbols, positions) where T

Construct a Cell type from arrays

CellBase.Cell — Method

Cell(lat::Lattice, numbers::Vector{Int}, positions)

Constructure the Cell type from lattice, positions and numbers

CellBase.array — Method

array(structure::Cell, arrayname::Symbol)

Return the additional array stored in the Cell object.

CellBase.arraynames — Method

arraynames(structure::Cell)

Return the names of additional arrays.

CellBase.atomic_numbers — Method

atomic_numbers(structure::Cell)

Return a Vector of the atomic numbers.

CellBase.attachmetadata! — Method

attachmetadata!(structure::Cell, metadata::Dict)

Replace metadata with an existing dictionary.

CellBase.cellmat — Method

cellmat(structure::Cell)

Return the matrix of lattice vectors.

CellBase.cellpar — Method

cellpar(structure::Cell)

Return the lattice parameters.

CellBase.check_minsep — Method

check_minsep(structure::Cell, minsep::Dict)

Check if the minimum separation constraints are satisfied. Minimum separations are supplied as an dictionary where the global version under the :global key. To supply the minimum separations between A and B, pass Dict((:A=>:B)=>1.0). Return true or false.

CellBase.clip — Method

clip(s::Cell, mask::AbstractVector)

Clip a structure with a given indexing array

CellBase.compute_minsep_mat — Method

minsep_matrix(structure::Cell, minsep::Dict)

Initialise the minimum separation matrix and species mapping. Returns the unique species, integer indexed species and the minimum separation matrix.

CellBase.distance_matrix — Method

distance_matrix(structure::Cell; mic=true)

Compute the distance matrix for the given structure, using the minimum image convention (or not).

Note the returned matrix does not expand the cell. The distance matrix cannot be safety used for obtaining the minimum separations. For example, a structure with a single atom would be a distance matrix containing only zero.

CellBase.distance_squared_between — Method

distance_squared_between(posmat::Matrix, i, j, svec::Matrix, ishift)

Return the squared distance between two positions stored in a matrix and shift vector.

CellBase.fingerprint — Method

fingerprint(s::Cell; dmat=distance_matrix(s), weighted=true, cut_bl=3.0)

Computed the fingerprint vector based on simple sorted pair-wise distances. NOTE: Does not work for single atom cell!!

CellBase.fingerprint_distance — Method

fingerprint_distance(f1::AbstractVector, f2::AbstractVector;lim=Inf)

Compute the deviation between two finger print vectors

Comparison is truncated by the size of the shortest vector of the two, or by the lim key word.

CellBase.formula_and_factor — Method

formula_and_factor(structure::Cell)

Return reduced formula and associated factor.

CellBase.get_cellmat — Method

get_cellmat(structure::Cell)

Return the matrix of lattice vectors(copy).

CellBase.get_lattice — Method

get_lattice(structure::Cell)

Return the Lattice instance (copy).

CellBase.get_positions — Method

get_positions(cell::Cell)

Return a copy of the positions (cartesian coordinates) of the atoms in a structure.

CellBase.get_scaled_positions — Method

get_fraction_positions(cell::Cell)

Return fractional positions of the cell.

CellBase.get_species — Method

get_species(structure::Cell)

Return a Vector (copy) of species names.

CellBase.lattice — Method

lattice(structure::Cell)

Return the Lattice instance.

CellBase.make_supercell — Method

make_supercell(structure::Cell, a, b, c)

Make a supercell

Currently only work with diagonal transform matrices. TODO: Write function for the general cases....

CellBase.metadata — Method

metadata(structure::Cell)

Return the metadata dictionary.

CellBase.natoms — Function

natoms(cell::Cell)

Return number of atoms in a structure.

CellBase.nions — Method

nions(cell::Cell)

Return number of atoms in a structure.

CellBase.num_fu — Method

num_fu(structure::Cell)

Return the number of formula units.

CellBase.positions — Method

positions(cell::Cell)

Return positions (cartesian coordinates) of the atoms in a structure.

CellBase.rattle! — Method

rattle!(cell::Cell, amp)

Rattle the positions of the cell for a given maximum amplitude (uniform distribution).

CellBase.reduced_fu — Method

reduced_fu(structure::Cell)

Return the reduced formula.

CellBase.set_cellmat! — Method

set_cellmat!(cell::Cell, mat;scale_positions=true)

Update the Lattice with a new matrix of lattice vectors. Scale of the existing postions if needed.

CellBase.set_positions! — Method

set_positions!(cell::Cell, pos)

Set the positions of the Cell with a new matrix.

CellBase.set_scaled_positions! — Method

set_scaled_positions!(cell::Cell, scaled::Matrix)

Set scaled positions for a cell.

CellBase.sorted_symbols — Method

sorted_symbols(symbols)

Return sorted symbols by atomic numbers.

CellBase.species — Method

species(structure::Cell)

Return a Vector of species names.

CellBase.specindex — Method

specindex(structure::Cell)

Return the unique species and integer based indices for each atom.

CellBase.sposarray — Method

sposarray(cell::Cell)

Return the positions as a Vector of static arrays. The returned array can provide improved performance for certain type of operations.

CellBase.volume — Method

volume(structure::Cell)

Return the volume of the cell.

CellBase.wrap! — Method

wrap!(cell::Cell)

Wrap atom outside of the lattice back into the box defined by the lattice vectors.

Spglib.jl interface

The routine in Spglib.jl (which wraps the spglib libraray) can be used for finding symmetry and performing reduction of the Cell type.

CellBase.Cell — Method

Cell(cell::SCell)

Return a Cell object from Spglib.Cell.

CellBase.SCell — Type

Alias for Spglib.Cell

CellBase.SCell — Method

SCell(cell::Cell)

Construct Spglib.Cell from Cell type.

CellBase.@extend_scell — Macro

Macro for extending the Spglib methods.

Usage:

@extend_scell get_dataset

will allow the get_dataset method of Spglib to be used for Cell type.

CellBase.@extend_scell_roundtrip — Macro

Macro for extending the Spglib methods and convert returned Spglib.Cell to Cell.

Usage:

@extend_scell_roundtrip standardize_cell

will allow the standardize_cell method to be used and the returned Spglib.Cell is converted to Cell.

Misc utils

Miscellaneous utility functions.

CellBase.reduction_gauss — Method

2D Gauss reduction

Works for column vectors instead of row vectors

CellBase.cellpar2mat — Method

convert cell parameters to column vectors

CellBase.isvalidcellpar — Method

Check if cell parameters are valid

CellBase.vec2cellpar — Method

Convert cell vectors to cell parameters

Returns an static array of the cell parameters

CellBase.volume — Method

Compute volume from cell parameters

Neighbour lists

Type and functions for handling neighbour lists.

CellBase.ExtendedPointArray — Type

Represent an array of points after expansion by periodic boundary

CellBase.ExtendedPointArray — Method

ExtendedPointArray(cell::Cell, rcut)

Constructed an ExtendedPointArray from a given structure

CellBase.NLIterator — Type

Iterator interface for going through all neighbours

CellBase.NLIteratorWithVector — Type

Iterator interface for going through all neighbours

CellBase.NeighbourList — Type

Type for representing a neighbour list

CellBase.NeighbourList — Type

NeighbourList(ea::ExtendedPointArray, rcut, nmax=100; savevec=false)

Construct a NeighbourList from an extended point array for the points in the original cell

CellBase.eachneighbour — Method

Iterate the neighbours of a site in the original cell. Returns a tuple of (originalindex, extendedindex, distance) for each iteration

CellBase.eachneighbourvector — Method

Iterate the neighbours of a site in the original cell. Returns a tuple of (originalindex, extendedindex, distance, vector) for each iteration

CellBase.nions_extended — Method

Number of ions in the extended cell

CellBase.nions_extended — Method

Number of ions in the extended cell

CellBase.nions_orig — Method

Number of ions in the original cell

CellBase.nions_orig — Method

Number of ions in the original cell

CellBase.num_neighbours — Method

Number of neighbours for a point

CellBase.rebuild! — Method

Rebuild the ExtendedPointArray for an existing cell

CellBase.rebuild! — Method

rebuild(nl::NeighbourList, cell::Cell)

Perform a full rebuild of the NeighbourList with the latest geometry of the cell

CellBase.rebuild! — Method

rebuild!(nl::NeighbourList, ea::ExtendedPointArray)

Perform a full rebuild of the neighbour list from scratch

CellBase.update! — Method

update!(nl::NeighbourList, cell::Cell)

Update the NeighbourList with the latest geometry of the Cell. No rebuilding is performed

CellBase.update! — Method

Update the calculated distances/vectors but do not rebuild the whole neighbour list

IO

Code for read/writing file of crystal structures.

CellBase.read_res — Method

Read an array containing the lines of the SHELX file

CellBase.write_res — Method

write_res(fname::AbstractString, structure::Cell)

Write out SHELX format data to a file

CellBase.write_res — Method

write_res(io::IO, structure::Cell)

Write out SHELX format data

CellBase.push_xyz! — Method

xyz lines for a single frame

CellBase.write_xyz — Method

Write snapshots to a xyz file

CellBase.CellIO._parse_tagline — Method

Parse the tag line, return a dictionary of the tagline is in the form of <ionsetname> % KEY=VALUE KEY=VALUE

CellBase.CellIO.clean_lines — Method

Clean lines by removing new line symbols and comments

CellBase.CellIO.filter_block — Method

Get rid of blocks

CellBase.CellIO.filter_block — Method

Filter away block names contained in only

CellBase.CellIO.find_block — Method

Find the block with given name, return a Vector of the lines Comments are skipped

CellBase.CellIO.parse_taglines — Method

Parse tag lines, return a vector of dictionary containing parsed tags

CellBase.CellIO.read_cell — Method

Read a cell file

CellBase.CellIO.read_cell — Method

Read content of a cell file

CellBase.CellIO.read_cellmat — Method

Read cell related sections

CellBase.CellIO.read_num_block — Method

Read a numerical block in the form of a vector of String

CellBase.CellIO.read_positions — Method

Read positions related sections

CellBase.CellIO.read_seed — Method

Read in cell

CellBase.CellIO.read_seed — Method

Read seed with the label parsed. At this stage no explansions are done

CellBase.DotCastep.SnapShot — Type

Representation of a snapshot

CellBase.DotCastep._get_nions — Method

Get the number of ions

CellBase.DotCastep._read_snapshot — Function

Read the initial structure and its energy/forces

CellBase.DotCastep.read_castep — Method

read_castep(fname::String)

Read a CASTEP file, return a Vector of the Snapshots

CellBase.DotCastep.read_coord_table! — Function

Read fractional coordinates (column vectors) from a table

CellBase.DotCastep.read_lattice — Method

Read lattice vectors from the lines (column vectors)

CellBase.DotCastep.skip_to_header — Method

Skip to the next header

CellBase.SheapIO.SheapOptions — Type

Commandline options for SHEAP

CellBase.SheapIO._SheapDefaultOptions — Type

Record the default options of SHEAP commandline

CellBase.SheapIO.parse_sheap_output — Method

Prase the output of SHEAP from an IO object

CellBase.SheapIO.run_sheap — Method

run_sheap(vecs, opt::SheapOptions;metadata=repeat([SheapMetadata()], length(vecs)), show_stderr=true)

Run SHEAP for a iterator of vectors with the given options.